week_08_multitask/hw_solution/requirements.txt
aiohttp
requests
Pillow
